requirements.txt
absl-py==0.11.0
astunparse==1.6.3
attrs==20.3.0
awscli==1.18.216
awsebcli==3.19.3
bcrypt==3.2.0
blessed==1.17.12
boto3==1.16.54
botocore==1.19.56
cached-property==1.5.2
cachetools==4.2.0
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==3.3.1
docker==4.4.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
flatbuffers==1.12
future==0.16.0
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.10.0
joblib==0.14.0
jsonschema==3.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.5
paramiko==2.7.2
pathspec==0.5.9
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.4
scipy==1.3.2
semantic-version==2.5.0
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
texttable==1.6.3
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==0.16.0
wrapt==1.12.1
xlrd==2.0.1
zipp==3.4.0
